+++ modules/apps/layout/layout-content-page-editor-web/tsconfig.json
@@ -15,10 +15,6 @@
 		"baseUrl": ".",
 		"composite": true,
 		"jsx": "react",
-		"lib": [
-			"DOM",
-			"es2019"
-		],
 		"module": "es6",
 		"moduleResolution": "node",
 		"outDir": "./types",
@@ -40,6 +36,10 @@
 		"typeRoots": [
 			"../../../node_modules/@types",
 			"./node_modules/@types"
+		],
+		"lib": [
+			"DOM",
+			"es2019"
 		]
 	},
 	"exclude": [
