+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/package.json
@@ -23,9 +23,6 @@
 		"react-router-dom": "^6.15.0",
 		"react-scripts": "5.0.1"
 	},
-	"devDependencies": {
-		"@liferay/npm-scripts": "^48.1.3"
-	},
 	"eslintConfig": {
 		"globals": {
 			"Liferay": true
@@ -36,8 +33,6 @@
 	"scripts": {
 		"build": "react-scripts build",
 		"eject": "react-scripts eject",
-		"format": "liferay-npm-scripts format --experimental-standalone 'src/**'",
-		"lint": "liferay-npm-scripts lint:fix --experimental-standalone 'src/**'",
 		"start": "react-scripts start",
 		"test": "react-scripts test --passWithNoTests --watchAll=false"
 	},
