+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/src/App.js
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+import {HashRouter, Route, Routes} from 'react-router-dom';
+
 import HomePage from './pages/HomePage/HomePage';
 import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
 import ProjectPage from './pages/ProjectPage/ProjectPage';
-
-import {HashRouter, Route, Routes} from 'react-router-dom';
 
 import './App.css';
 
@@ -15,14 +15,14 @@ function App() {
 	return (
 		<HashRouter>
 			<Routes>
-				<Route path="/" element={<HomePage />} />
+				<Route element={<HomePage />} path="/" />
 				<Route path="/projects">
-					<Route path=":id" element={<ProjectPage />} />
+					<Route element={<ProjectPage />} path=":id" />
 				</Route>
-				<Route path="*" element={<NotFoundPage />} />
+				<Route element={<NotFoundPage />} path="*" />
 			</Routes>
 		</HashRouter>
 	);
-};
+}
 
 export default App;
