+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/src/components/Breadcrumbs/Breadcrumbs.js
@@ -7,9 +7,9 @@ import {Link} from 'react-router-dom';
 
 function Breadcrumbs({breadcrumbs}) {
 	return (
-		<ol class="breadcrumb">
-			{breadcrumbs && breadcrumbs.map(
-				(breadcrumb) => {
+		<ol className="breadcrumb">
+			{breadcrumbs &&
+				breadcrumbs.map((breadcrumb) => {
 					if (breadcrumb.active) {
 						return (
 							<li className="breadcrumb-item">
@@ -21,7 +21,7 @@ function Breadcrumbs({breadcrumbs}) {
 					}
 
 					return (
-						<li className="breadcrumb-item">
+						<li className="breadcrumb-item" key={breadcrumb.name}>
 							<Link
 								className="breadcrumb-link"
 								title={breadcrumb.name}
@@ -33,8 +33,7 @@ function Breadcrumbs({breadcrumbs}) {
 							</Link>
 						</li>
 					);
-				}
-			)}
+				})}
 		</ol>
 	);
 }
