+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/src/pages/HomePage/HomePage.js
@@ -4,24 +4,23 @@
  */
 
 import ClayCard from '@clayui/card';
-import ClayLayout from '@clayui/layout';
 import {Heading} from '@clayui/core';
+import ClayLayout from '@clayui/layout';
 
 import './HomePage.css';
-
 import Breadcrumbs from '../../components/Breadcrumbs/Breadcrumbs';
 import ProjectQueue from '../../components/ProjectQueue/ProjectQueue';
 
 function Home() {
-	const breadcrumbs = [
-		{active:true, link: '/', name: 'Home'},
-	];
+	const breadcrumbs = [{active: true, link: '/', name: 'Home'}];
 
 	return (
 		<ClayLayout.Container>
 			<ClayCard className="jethr0-card">
 				<Breadcrumbs breadcrumbs={breadcrumbs} />
-				<Heading level={3} weight="lighter">Project Queue</Heading>
+				<Heading level={3} weight="lighter">
+					Project Queue
+				</Heading>
 				<ProjectQueue />
 			</ClayCard>
 		</ClayLayout.Container>
