+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-custom-element/src/pages/NotFoundPage/NotFoundPage.js
@@ -4,9 +4,7 @@
  */
 
 function NotFoundPage() {
-	return (
-		<div>404 - Not Found</div>
-	);
+	return <div>404 - Not Found</div>;
 }
 
 export default NotFoundPage;
